editVideo & generateVideo methods done, interfaces updated

+++ src/index.ts
@@ -1,27 +1,38 @@
 import InstagramScraper from './providers/instagram';
 import { PuppeteerLaunchOptions } from 'puppeteer';
 import { GenerationOptions } from './types/generationOptions';
-import { InstagramPost } from './types/instagramPost';
 
 import ffmpeg from 'fluent-ffmpeg'
 import ffmpegPath from '@ffmpeg-installer/ffmpeg'
 import ffprobePath from '@ffprobe-installer/ffprobe';
 import path from 'path';
 import { v4 as uuidv4 } from 'uuid';
-import fs, { constants } from 'fs';
+import fs from 'fs';
 
-interface BrainrotGeneratorOptions {
+interface ShortsGeneratorOptions {
+    outputDir?: string;
     backgroundVideosDir?: string;
     puppeteerLaunchOptions?: PuppeteerLaunchOptions;
 }
 
-class BrainrotGenerator extends InstagramScraper {
-    private accountsSource: string;
-    private options: BrainrotGeneratorOptions;
+class ShortsGenerator extends InstagramScraper {
+    private accountsSource: string[];
+    private options: ShortsGeneratorOptions;
 
-    constructor(accountsSource: string, options?: BrainrotGeneratorOptions) {
+    constructor(accountsSource: string[], options?: ShortsGeneratorOptions) {
         super(options?.puppeteerLaunchOptions)
+
         this.options = options || {};
+
+
+        if(!options?.outputDir) {
+            this.options['outputDir'] = __dirname
+        }
+
+        if(!options?.backgroundVideosDir) {
+            this.options['backgroundVideosDir'] = path.join(__dirname, '/bg_videos')
+        }
+
         this.accountsSource = accountsSource;
 
 
@@ -42,11 +53,12 @@ class BrainrotGenerator extends InstagramScraper {
 
             let account = options.account || this.getRandomAccount()
 
-            const posts: InstagramPost[] = await this.scrapeUsersPosts(account, options.count);
-
-            for await (let post of posts) {
-                // await this.editVideo(post.videoUrl, )
-            }
+            const sourceVideo = options.sourceVideo || (await this.scrapeUsersPosts(account, 1))[0].videoUrl;
+            
+            const backgroundVideo = options.backgroundVideo;
+            if(!backgroundVideo) throw new Error('No video provided')
+            
+            await this.editVideo(sourceVideo, options.backgroundVideo)
 
         } catch (e) {
             console.log('generateVideo -> catch', e);
@@ -54,18 +66,16 @@ class BrainrotGenerator extends InstagramScraper {
         }
     }
 
-    private async createDirectoryIfNotExists(dir: string): Promise<string> {
+    private async createDirectoryIfNotExists(dirPath: string): Promise<string> {
         try {
 
-            const finalPath = path.join(__dirname, dir) 
-
-            if (!fs.existsSync(finalPath)) {
-                fs.mkdirSync(finalPath, {
+            if (!fs.existsSync(dirPath)) {
+                fs.mkdirSync(dirPath, {
                     recursive: true
                 })
             }
             
-            return finalPath
+            return dirPath
 
         } catch (e) {
             throw e
@@ -73,18 +83,16 @@ class BrainrotGenerator extends InstagramScraper {
     }
 
 
-    private async createFileIfNotExists(filename: string): Promise<string> {
+    private async createFileIfNotExists(filePath: string): Promise<string> {
         try {
 
-            const finalPath = path.join(__dirname, filename) 
-
             try {
-                fs.accessSync(finalPath, fs.constants.F_OK)
+                fs.accessSync(filePath, fs.constants.F_OK)
             } catch(e) {
-                fs.writeFileSync(finalPath, '')
+                fs.writeFileSync(filePath, '')
             }
             
-            return finalPath
+            return filePath
             
         } catch (e) {
             throw e
@@ -94,16 +102,26 @@ class BrainrotGenerator extends InstagramScraper {
     private async deleteUnprocessedVideos(outputDir: string, list: string[]): Promise<string[]> {
         try {
 
-            const files = fs.readdirSync(outputDir)
-            
-            // Determine files to delete
-            const filesToDelete = files.filter(file => !list.includes(file));
+            // Helper function to extract UUID from the filename
+            function extractUUID(filename: string): string | null {
+                const match = filename.match(/_([a-f0-9\-]{36})_/);
+                return match ? match[1] : null;
+            }
 
-            // Delete each file
-            filesToDelete.forEach(file => {
-                const filePath = path.join(outputDir, file);
-                fs.unlinkSync(filePath);
-            });
+            const files = fs.readdirSync(outputDir)
+            const uuid = extractUUID(list[0]) || null
+
+            if(uuid) {
+                // Determine files to delete
+                const filesToDelete = files.filter(file => !list.includes(file) && file.includes(uuid));
+    
+                // Delete each file
+                filesToDelete.forEach(file => {
+                    const filePath = path.join(outputDir, file);
+                    fs.unlinkSync(filePath);
+                });
+    
+            }
 
             return list
 
@@ -121,8 +139,8 @@ class BrainrotGenerator extends InstagramScraper {
                 return videosNames
             }
 
-            const outputDir = await this.createDirectoryIfNotExists(this.options?.backgroundVideosDir || './output/bg_videos/');
-            const listFile = await this.createFileIfNotExists('./list.txt')
+            const outputDir = await this.createDirectoryIfNotExists(this.options.backgroundVideosDir!);
+            const listFile = await this.createFileIfNotExists(path.join(this.options.backgroundVideosDir!, './list.txt'))
 
             const filePattern = path.join(outputDir, `bg_video_${uuidv4()}_%03d.mp4`);
 
@@ -137,8 +155,8 @@ class BrainrotGenerator extends InstagramScraper {
                 ];
     
                 const ffmpegProcess = ffmpeg(videoUrl)
-                    .videoFilters('crop=ih*9/8:ih')
-                    .size('1080x960')
+                    .videoFilters('crop=ih*9/16:ih')
+                    .size('1080x1920')
                     .outputOptions(outputOptions)
                     .audioCodec('aac') 
                     .audioBitrate('192k')
@@ -179,7 +197,45 @@ class BrainrotGenerator extends InstagramScraper {
     // Todo: check if sourceVideo & backgroundVideo should be path / buffer or any.
     private async editVideo(sourceVideo: any, backgroundVideo: any): Promise<void> {
         try {
-            return;
+            let output = path.join(this.options.outputDir!, `edited_${uuidv4()}.mp4`)
+            await new Promise<void>((resolve, reject) => {
+
+                ffmpeg.ffprobe(sourceVideo, (e, metadata) => {
+                    if (e) reject(e);
+                
+                    const duration = metadata.format.duration!;
+                    const width = Math.min(1000, metadata.streams[0].width!);
+
+                    const fadeOutStart = (duration - 1).toFixed(0);
+                    const cropDuration = (duration + 3).toFixed(0);
+                
+                    ffmpeg()
+                    .input(backgroundVideo)
+                    .input(sourceVideo)
+                    .complexFilter([
+                        // Scale the background video to 1080x1920
+                        '[0:v]scale=1080:1920[bg]',
+                        // Scale the source video to 95% of the background video size
+                        `[1:v]scale=${width}:-1[src]`,
+                        // Apply fade-in and fade-out effects to the source video
+                        '[src]format=yuva420p,fade=t=in:st=1:d=1:alpha=1,fade=t=out:st=' + fadeOutStart + ':d=1:alpha=1[trans]',
+                        // Overlay the source video on the background video, centered
+                        '[bg][trans]overlay=(W-w)/2:(H-h)/2[outv]'
+                    ])
+                    .outputOptions([
+                        '-map [outv]',   // Output the combined video stream
+                        '-map 1:a',      // Map the audio from the source video
+                    ])
+                    .output(output)
+                    .duration(cropDuration) // Crop the video to the desired duration
+                    .on("error", () => {})
+                    .on("end", () => {
+                        resolve();
+                    })
+                    .run();
+                });
+            })
+
         } catch (e) {
             console.log('editVideo -> catch', e);
             throw e;
@@ -188,4 +244,4 @@ class BrainrotGenerator extends InstagramScraper {
 
 }
 
-export default BrainrotGenerator
+export default ShortsGenerator
